Added optuna + started troubleshooting
TODO: need to fix dataloaders. I think it is best
to start with data being flattened as required
by the model, then create a new function which
converts any given tensor to the correct image
shape.
Then, should also ensure output is in the correct
form for an image, eg between 0 and 255?